Application ported to SDL

+++ CoffeeEngine/src/CoffeeEngine/Core/Application.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "Base.h"
+#include "SDL3/SDL_stdinc.h"
 #include "Window.h"
 #include "LayerStack.h"
 #include "CoffeeEngine/Events/ApplicationEvent.h"
@@ -89,7 +90,7 @@ namespace Coffee
         ImGuiLayer* m_ImGuiLayer; ///< The ImGui layer.
         bool m_Running = true; ///< Indicates whether the application is running.
         LayerStack m_LayerStack; ///< The stack of layers.
-        float m_LastFrameTime = 0.0f; ///< The time of the last frame.
+        Uint64 m_LastFrameTime = 0.0f; ///< The time of the last frame.
 
       private:
         static Application* s_Instance; ///< The singleton instance of the Application.
